(refactor) Various fixes and improvements
