+++ .husky/pre-push
@@ -3,5 +3,4 @@
 
 set -e  # die on error
 
-yarn run verify
-
+yarn verify
