+++ package.json
@@ -11,11 +11,11 @@
     "serve": "webpack serve --mode=development",
     "build": "webpack --mode production",
     "analyze": "webpack --mode=production --env.analyze=true",
-    "lint": "eslint \"src/**/*.ts\" \"src/**/*.tsx\" --fix --max-warnings=0 -c .eslintrc.js",
+    "lint": "TIMING=0 eslint \"src/**/*.ts\" \"src/**/*.tsx\" --fix --max-warnings=0",
     "prettier": "prettier --write \"src/**/*.{ts,tsx}\"",
     "typescript": "tsc",
     "test": "jest --config jest.config.json --passWithNoTests",
-    "verify": "concurrently 'yarn:lint' 'yarn:test' 'yarn:typescript'",
+    "verify": "turbo lint typescript coverage",
     "coverage": "yarn test -- --coverage ",
     "postinstall": "husky install",
     "extract-translations": "i18next 'src/**/*.tsx' --config ./tools/i18next-parser.config.js"
@@ -48,52 +48,54 @@
     "react-router-dom": "6.x"
   },
   "devDependencies": {
-    "@carbon/react": "^1.9.0",
+    "@carbon/react": "^1.37.0",
     "@openmrs/esm-framework": "next",
-    "@swc-node/loader": "^1.3.5",
-    "@swc/core": "^1.2.245",
-    "@swc/jest": "^0.2.22",
-    "@testing-library/dom": "^7.20.0",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.3.0",
+    "@swc-node/loader": "^1.3.7",
+    "@swc/core": "^1.3.83",
+    "@swc/jest": "^0.2.29",
+    "@testing-library/dom": "^7.31.2",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^28.1.7",
-    "@types/react-dom": "^18.0.6",
+    "@types/jest": "^28.1.8",
+    "@types/react-dom": "^18.2.7",
     "@types/react-router-dom": "^5.3.3",
-    "@types/testing-library__jest-dom": "^5.14.5",
-    "@types/webpack-env": "^1.16.0",
-    "@typescript-eslint/parser": "^5.14.0",
-    "concurrently": "^6.2.0",
+    "@types/testing-library__jest-dom": "^5.14.9",
+    "@types/webpack-env": "^1.18.1",
+    "@typescript-eslint/eslint-plugin": "^6.6.0",
+    "@typescript-eslint/parser": "^6.6.0",
+    "concurrently": "^6.5.1",
     "css-loader": "^6.8.1",
-    "eslint": "^8.20.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint": "^8.48.0",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-config-react-app": "^7.0.1",
     "eslint-config-ts-react-important-stuff": "^3.0.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "husky": "^8.0.1",
+    "husky": "^8.0.3",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^28.1.3",
     "jest-cli": "^28.1.3",
     "jest-environment-jsdom": "^28.1.3",
     "lodash-es": "^4.17.21",
     "openmrs": "next",
-    "prettier": "^2.3.0",
-    "pretty-quick": "^3.1.0",
+    "prettier": "^2.8.8",
+    "pretty-quick": "^3.1.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-i18next": "^11.18.4",
-    "react-router-dom": "^6.3.0",
-    "semver": "^7.3.7",
+    "react-i18next": "^11.18.6",
+    "react-router-dom": "^6.15.0",
+    "semver": "^7.5.4",
     "swc-loader": "^0.2.3",
-    "typescript": "^4.7.3",
-    "webpack": "^5.73.0",
+    "turbo": "^1.10.13",
+    "typescript": "^4.9.5",
+    "webpack": "^5.88.2",
     "webpack-cli": "^5.1.4"
   },
   "packageManager": "yarn@3.2.2",
   "dependencies": {
     "i18next": "^21.10.0",
     "i18next-parser": "^6.6.0",
-    "react-hook-form": "^7.34.2",
-    "swr": "^2.1.3"
+    "react-hook-form": "^7.46.1",
+    "swr": "^2.2.2"
   }
 }
