+++ src/add-group-modal/styles.scss
@@ -1,5 +1,5 @@
-@use '@carbon/styles/scss/spacing';
 @use '@carbon/colors';
+@use '@carbon/layout';
 
 .modal {
   :global(.cds--modal) {
@@ -32,7 +32,7 @@
 
 .patientName {
   flex-grow: 1;
-  padding-left: spacing.$spacing-05;
+  padding-left: layout.$spacing-05;
 }
 
 .loading {
@@ -41,5 +41,5 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  row-gap: spacing.$spacing-05;
+  row-gap: layout.$spacing-05;
 }
