+++ src/context/GroupFormWorkflowReducer.ts
@@ -205,7 +205,6 @@ const reducer = (state, action) => {
           activeFormUuid: null,
         };
         persistData(newState);
-        // eslint-disable-next-line
         navigate({ to: "${openmrsSpaBase}/forms" });
         return newState;
       } else if (thisForm.workflowState === "SUBMIT_FOR_NEXT") {
@@ -408,7 +407,6 @@ const reducer = (state, action) => {
         activeFormUuid: null,
       };
       persistData(newState);
-      //eslint-disable-next-line
       navigate({ to: "${openmrsSpaBase}/forms" });
       return { ...newState, formDestroyed: true };
     }
@@ -418,7 +416,6 @@ const reducer = (state, action) => {
         activeFormUuid: null,
       };
       persistData(newState);
-      //eslint-disable-next-line
       navigate({ to: "${openmrsSpaBase}/forms" });
       return newState;
     }
