+++ src/declarations.d.ts
@@ -1,0 +1,8 @@
+declare module "@carbon/react";
+declare module "*.css";
+declare module "*.scss" {
+  const content: { [className: string]: string };
+  export = content;
+}
+
+declare type SideNavProps = object;
