--- src/declarations.d.tsx
@@ -1,2 +1,0 @@
-declare module "*.css";
-declare module "*.scss";
