+++ src/form-entry-workflow/form-review-card/styles.scss
@@ -1,12 +1,10 @@
-@use '@carbon/styles/scss/spacing';
-// @use '@carbon/colors';
-@use '@carbon/styles/scss/type';
-@import '~@openmrs/esm-styleguide/src/vars';
-
+@use '@carbon/colors';
+@use '@carbon/layout';
+@use '@carbon/type';
 
 .formReviewCard {
-  background-color: $ui-02;
-  padding: spacing.$spacing-02;
+  background-color: colors.$white-0;
+  padding: layout.$spacing-02;
 }
 
 .formReviewCard :global(.cds--accordion) :global(.cds--accordion__item) {
@@ -16,17 +14,17 @@
 .formReviewCard :global(.cds--accordion__title) {
   display: flex;
   align-items: baseline;
-  column-gap: spacing.$spacing-05;
+  column-gap: layout.$spacing-05;
 }
 
 .formReviewCard :global(.cds--accordion__content) {
-  padding: spacing.$spacing-03;
+  padding: layout.$spacing-03;
 }
 
 .dataField {
   @include type.type-style('code-02');
-  background-color: $ui-01;
-  padding: spacing.$spacing-03;
+  background-color: colors.$gray-10;
+  padding: layout.$spacing-03;
 }
 
 .displayName {
