+++ src/form-entry-workflow/patient-banner/styles.scss
@@ -1,16 +1,15 @@
-@use '@carbon/styles/scss/spacing';
-// @use '@carbon/colors';
-@use '@carbon/styles/scss/type';
-@import '~@openmrs/esm-styleguide/src/vars';
+@use '@carbon/colors';
+@use '@carbon/layout';
+@use '@carbon/type';
 
 .container {
-  height: spacing.$spacing-11;
+  height: layout.$spacing-11;
   display: flex;
   align-items: center;
-  background-color: $ui-02;
-  border-top: 0.0125rem solid $ui-03;
-  border-bottom: 0.0125rem solid $ui-03;
-  padding: 0 spacing.$spacing-05;
+  background-color: colors.$white-0;
+  border-top: 0.0125rem solid colors.$gray-20;
+  border-bottom: 0.0125rem solid colors.$gray-20;
+  padding: 0 layout.$spacing-05;
 }
 
 .photoPlaceholder {
@@ -35,11 +34,11 @@
     min-height: 2rem;
   }
   @include type.type-style('body-compact-02');
-  color: $text-02;
+  color: colors.$gray-70;
   column-gap: 0.8rem;
 }
 
 .patientEditBtn {
-  color: $ui-05;
-  margin: spacing.$spacing-03;
+  color: colors.$gray-100;
+  margin: layout.$spacing-03;
 }
