+++ src/form-entry-workflow/workflow-review/styles.scss
@@ -1,7 +1,3 @@
-@import "~@openmrs/esm-styleguide/src/vars";
-@import "~carbon-components/src/globals/scss/vars";
-@import "~carbon-components/src/globals/scss/mixins";
-
 .workspaceWrapper {
   display: flex;
   justify-content: center;
