+++ src/forms-app-menu-link.tsx
@@ -5,7 +5,6 @@ import { ConfigurableLink } from "@openmrs/esm-framework";
 export default function FormsAppMenuLink() {
   const { t } = useTranslation();
   return (
-    // eslint-disable-next-line
     <ConfigurableLink to="${openmrsSpaBase}/forms">
       {t("formsAppMenuLink", "Fast Data Entry")}
     </ConfigurableLink>
