+++ src/group-form-entry-workflow/group-search/group-search.scss
@@ -1,17 +1,15 @@
-@use '@carbon/styles/scss/spacing';
-@use '@carbon/styles/scss/type';
-@import '~@openmrs/esm-styleguide/src/vars';
-
-
+@use '@carbon/colors';
+@use '@carbon/layout';
+@use '@carbon/type';
 
 .searchResultsContainer {
   width: 100%;
-  background-color: $ui-02;
+  background-color: colors.$white-0;
 
   a {
     text-decoration: none;
     @include type.type-style('heading-02');
-    color: $text-02;
+    color: colors.$gray-70;
     margin: 0rem;
   }
 }
@@ -22,7 +20,7 @@
 
 :global(.omrs-breakpoint-gt-tablet) .searchResultsContainer {
   padding: 0;
-  top: spacing.$spacing-09;
+  top: layout.$spacing-09;
 }
 
 .searchResults {
@@ -38,9 +36,9 @@
 
 .resultsText {
   @include type.type-style('label-01');
-  color: $text-02;
-  line-height: spacing.$spacing-05;
-  margin: spacing.$spacing-03 spacing.$spacing-05;
+  color: colors.$gray-70;
+  line-height: layout.$spacing-05;
+  margin: layout.$spacing-03 layout.$spacing-05;
 }
 
 .helperText {
@@ -50,46 +48,46 @@
 
 .emptyResultText {
   @include type.type-style('heading-compact-01');
-  color: $text-02;
-  margin-top: spacing.$spacing-05;
+  color: colors.$gray-70;
+  margin-top: layout.$spacing-05;
   margin-bottom: 0.313rem;
 }
 
 .actionText {
   @include type.type-style('body-01');
-  color: $text-02;
+  color: colors.$gray-70;
 }
 
 .pagination {
   display: flex;
   justify-content: space-evenly;
-  padding: 4.688rem 0 spacing.$spacing-06;
+  padding: 4.688rem 0 layout.$spacing-06;
 }
 
 .emptySearchResultsTile {
   text-align: center;
-  margin-top: spacing.$spacing-05;
-  padding: spacing.$spacing-09 0rem;
+  margin-top: layout.$spacing-05;
+  padding: layout.$spacing-09 0rem;
 }
 
 :global(.omrs-breakpoint-gt-tablet) .emptySearchResultsTile {
-  margin: spacing.$spacing-05;
+  margin: layout.$spacing-05;
 }
 
 .errorMessage {
   @include type.type-style('heading-compact-02');
   margin-top: 2.25rem;
-  margin-bottom: spacing.$spacing-03;
+  margin-bottom: layout.$spacing-03;
 }
 
 .errorCopy {
-  margin-bottom: spacing.$spacing-03;
+  margin-bottom: layout.$spacing-03;
   @include type.type-style('body-01');
-  color: $text-02;
+  color: colors.$gray-70;
 }
 
 .lastItem {
-  padding: spacing.$spacing-05;
+  padding: layout.$spacing-05;
   display: flex;
   justify-content: center;
   align-items: center;
