+++ src/index.ts
@@ -19,7 +19,7 @@ export const importTranslation = require.context(
   "lazy"
 );
 
-export const root = getAsyncLifecycle(() => import("./Root"), options);
+export const root = getAsyncLifecycle(() => import("./root"), options);
 
 export const formsAppMenuLink = getAsyncLifecycle(
   () => import("./forms-app-menu-link"),
